+++ recipes/default.rb
@@ -7,37 +7,19 @@
 # Distributed under MIT license
 #
 
-# add CUBRID Launchpad PPA
-#apt_repository "cubrid" do
-#  uri "http://ppa.launchpad.net/cubrid/cubrid/ubuntu/"
-  # for this case, use lucid support only,
-  # later it should auto detect the user OS
-#  components ["lucid", "main"]
-#  keyserver "keyserver.ubuntu.com"
-#  key "E871ADEE"
-#end
-
-FILENAME = "#{node['cubrid']['filename']}"
+FILENAME = node['cubrid']['filename']
 TEMP_DIR = "/tmp"
 CUBRID_BINARY = "#{TEMP_DIR}/#{FILENAME}"
-CUBRID_HOME_DIR = "#{node['cubrid']['home']}"
-CUBRID_DATABASES_DIR = "#{CUBRID_HOME_DIR}/databases"
-CUBRID_CONF = "#{node['cubrid']['conf']}"
-CM_HTTPD_CONF = "#{node['cubrid']['cm_httpd_conf']}"
-ENV_SCRIPT = "#{node['cubrid']['env_script']}"
-
-ENV['CUBRID'] = CUBRID_HOME_DIR
-ENV['CUBRID_DATABASES'] = CUBRID_DATABASES_DIR
-ENV['CUBRID_LANG'] = node['cubrid']['lang']
-ENV['CUBRID_CHARSET'] = node['cubrid']['charset']
-ENV['LD_LIBRARY_PATH'] = "#{CUBRID_HOME_DIR}/lib:#{ENV['LD_LIBRARY_PATH']}"
-ENV['PATH'] = "#{CUBRID_HOME_DIR}/bin:#{ENV['PATH']}"
+CUBRID_HOME_DIR = node['cubrid']['home']
+CUBRID_CONF = node['cubrid']['conf']
+CM_HTTPD_CONF = node['cubrid']['cm_httpd_conf']
+ENV_SCRIPT = node['cubrid']['env_script']
 BROKER_CONF = node['cubrid']['broker_conf']
 
 # Download the archive.
 remote_file CUBRID_BINARY do
   action :create_if_missing
-  source "#{node['cubrid']['tar_url']}"
+  source node['cubrid']['download_url']
   mode 0644
   not_if "test -d #{CUBRID_HOME_DIR}"
 end
@@ -56,14 +38,13 @@ execute "mv CUBRID #{CUBRID_HOME_DIR}" do
 end
 
 # Own CUBRID files.
-execute "Change ownership of #{CUBRID_HOME_DIR}" do
-  command "chown -R root #{CUBRID_HOME_DIR} && chgrp -R root #{CUBRID_HOME_DIR}"
+execute "chown -R root:root #{CUBRID_HOME_DIR}" do
   only_if "test -d #{CUBRID_HOME_DIR}"
   not_if "ls -ld #{CUBRID_HOME_DIR} | grep root"
 end
 
 # Remove the downloaded archive.
-file "#{CUBRID_BINARY}" do
+file CUBRID_BINARY do
   action :delete
   backup false
   only_if "test -f #{CUBRID_BINARY}"
@@ -75,7 +56,7 @@ template ENV_SCRIPT do
   not_if "cat #{ENV_SCRIPT} | grep 'Cookbook Name:: cubrid'"
 end
 
-# update cubrid.conf
+# Update cubrid.conf.
 template CUBRID_CONF do
   source "default.cubrid.conf.erb"
   not_if "cat #{CUBRID_CONF} | grep 'Cookbook Name:: cubrid'"
@@ -97,8 +78,15 @@ template CM_HTTPD_CONF do
   not_if "cat #{CM_HTTPD_CONF} | grep 'Cookbook Name:: cubrid'"
 end
 
-# Start CUBRID Service.
-execute "cubrid service start"
+ENV['CUBRID'] = CUBRID_HOME_DIR
+ENV['CUBRID_DATABASES'] = node['cubrid']['databases_dir']
+ENV['CUBRID_LANG'] = node['cubrid']['lang']
+ENV['CUBRID_CHARSET'] = node['cubrid']['charset']
+ENV['LD_LIBRARY_PATH'] = "#{CUBRID_HOME_DIR}/lib:#{ENV['LD_LIBRARY_PATH']}"
+ENV['PATH'] = "#{CUBRID_HOME_DIR}/bin:#{ENV['PATH']}"
+
+# Retart CUBRID Service.
+execute "cubrid service restart"
 
 # On CentOS/RedHat/Fedora iptables REJECTs all external connection to most ports including those used to conenct to CUBRID.
 # We need to open CUBRID-only ports.